Renames file to async_supabase_vector_store to match class name
